build block works with a OrdersSource
