+++ crates/rbuilder/src/backtest/execute.rs
@@ -18,6 +18,8 @@ use reth_provider::{BlockReader, DatabaseProviderFactory, HeaderProvider, StateP
 use serde::{Deserialize, Serialize};
 use std::{cell::RefCell, rc::Rc, sync::Arc};
 
+use super::OrdersWithTimestamp;
+
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
 pub struct BacktestBuilderOutput {
     pub orders_included: usize,
@@ -60,7 +62,6 @@ pub fn backtest_prepare_ctx_for_block<P>(
     block_data: BlockData,
     provider: P,
     chain_spec: Arc<ChainSpec>,
-    build_block_lag_ms: i64,
     blocklist: HashSet<Address>,
     sbundle_mergeabe_signers: &[Address],
     builder_signer: Signer,
@@ -68,18 +69,6 @@ pub fn backtest_prepare_ctx_for_block<P>(
 where
     P: StateProviderFactory + Clone + 'static,
 {
-    let orders = block_data
-        .available_orders
-        .iter()
-        .filter_map(|order| {
-            if order.timestamp_ms as i64 + build_block_lag_ms
-                >= block_data.winning_bid_trace.timestamp_ms as i64
-            {
-                return None;
-            }
-            Some(order.order.clone())
-        })
-        .collect::<Vec<_>>();
     let ctx = BlockBuildingContext::from_onchain_block(
         block_data.onchain_block,
         chain_spec.clone(),
@@ -89,8 +78,30 @@ where
         block_data.winning_bid_trace.proposer_fee_recipient,
         Some(builder_signer),
     );
+    backtest_prepare_ctx_for_block_from_building_context(
+        ctx,
+        block_data.available_orders,
+        provider,
+        sbundle_mergeabe_signers,
+    )
+}
+
+pub fn backtest_prepare_ctx_for_block_from_building_context<P>(
+    ctx: BlockBuildingContext,
+    available_orders: Vec<OrdersWithTimestamp>,
+    provider: P,
+    sbundle_mergeabe_signers: &[Address],
+) -> eyre::Result<BacktestBlockInput>
+where
+    P: StateProviderFactory + Clone + 'static,
+{
+    let orders = available_orders
+        .iter()
+        .map(|order| order.order.clone())
+        .collect::<Vec<_>>();
+
     let (sim_orders, sim_errors) =
-        simulate_all_orders_with_sim_tree(provider.clone(), &ctx, &orders, false)?;
+        simulate_all_orders_with_sim_tree(provider, &ctx, &orders, false)?;
 
     // Apply bundle merging as in live building.
     let order_store = Rc::new(RefCell::new(SimulatedOrderStore::new()));
@@ -111,7 +122,6 @@ pub fn backtest_simulate_block<P, DB, ConfigType>(
     block_data: BlockData,
     provider: P,
     chain_spec: Arc<ChainSpec>,
-    build_block_lag_ms: i64,
     builders_names: Vec<String>,
     config: &ConfigType,
     blocklist: HashSet<Address>,
@@ -134,7 +144,6 @@ where
         block_data.clone(),
         provider.clone(),
         chain_spec.clone(),
-        build_block_lag_ms,
         blocklist,
         sbundle_mergeabe_signers,
         config.base_config().coinbase_signer()?,
