+++ crates/rbuilder/src/backtest/mod.rs
@@ -1,14 +1,13 @@
-mod backtest_build_block;
 mod backtest_build_range;
 pub mod execute;
 pub mod fetch;
 
+pub mod build_block;
 pub mod redistribute;
 pub mod restore_landed_orders;
 mod results_store;
 mod store;
 
-pub use backtest_build_block::run_backtest_build_block;
 pub use backtest_build_range::run_backtest_build_range;
 use std::collections::HashSet;
