+++ crates/rbuilder/src/bin/backtest-build-block.rs
@@ -1,8 +1,10 @@
 //! Instantiation of run_backtest_build_block on our sample configuration.
 
-use rbuilder::{backtest::run_backtest_build_block, live_builder::config::Config};
+use rbuilder::{
+    backtest::build_block::landed_block_from_db::run_backtest, live_builder::config::Config,
+};
 
 #[tokio::main]
 async fn main() -> eyre::Result<()> {
-    run_backtest_build_block::<Config>().await
+    run_backtest::<Config>().await
 }
